moved urls to compute package and tls config to node package
