+++ node/config.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/ChristianMct/helium"
 	"github.com/ChristianMct/helium/objectstore"
 	"github.com/ChristianMct/helium/services/compute"
 	"github.com/ChristianMct/helium/services/setup"
@@ -25,7 +24,7 @@ type Config struct {
 	SetupConfig       setup.ServiceConfig
 	ComputeConfig     compute.ServiceConfig
 	ObjectStoreConfig objectstore.Config
-	TLSConfig         helium.TLSConfig
+	TLSConfig         TLSConfig
 }
 
 // Address is the network address of a node.
@@ -103,4 +102,19 @@ func ValidateConfig(config Config, nl List) error {
 		return fmt.Errorf("helper ID `%s` not found in the node list", config.HelperID)
 	}
 	return nil
+}
+
+// TLSConfig is a struct for specifying TLS-related configuration.
+// TLS is not supported yet.
+//
+//nolint:gosec // sha1 needed to check certificate
+type TLSConfig struct {
+	InsecureChannels bool                      // if set, disables TLS authentication
+	FromDirectory    string                    // path to a directory containing the TLS material as files
+	PeerPKs          map[session.NodeID]string // Mapping of <node, pubKey> where pubKey is PEM encoded
+	PeerCerts        map[session.NodeID]string // Mapping of <node, certifcate> where pubKey is PEM encoded ASN.1 DER string
+	CACert           string                    // Root CA certificate as a PEM encoded ASN.1 DER string
+	OwnCert          string                    // Own certificate as a PEM encoded ASN.1 DER string
+	OwnPk            string                    // Own public key as a PEM encoded string
+	OwnSk            string                    // Own secret key as a PEM encoded string
 }
